backend/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.15.2
django-filter==25.2
django-cors-headers==4.4.0
dj-database-url==2.1.0
pillow==12.0.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-decouple==3.8
python-dotenv==1.0.0
django-storages[boto3]==1.14.2
sqlparse==0.5.4
tzdata==2025.3
boto3==1.35.46
botocore==1.34.46
yoco==1.2.0
requests==2.31.0


